id_sd/in/vl: fix remaining Uint32 SDL_GetTicks truncation instances
Missed in the previous pass: id_sd.c SD_Poll music/sfx timing locals,
id_in.c IN_UserInput target deadline, id_vl.c VW_UpdateScreen frame-
pacing 'now'. All stored SDL_GetTicks() (Uint64) in Uint32 variables;
updated to Uint64. SDL_Delay takes Uint32 so the pacing delta cast is
made explicit.

+++ src/id_in.c
@@ -597,8 +597,8 @@ void IN_SetControlType(int player, ControlType type)
 boolean IN_UserInput(longword delay)
 {
     // delay is in 70Hz ticks; convert to milliseconds for SDL_GetTicks
-    Uint32 ms = (Uint32)delay * 1000u / 70u;
-    Uint32 target = SDL_GetTicks() + ms;
+    Uint64 ms = (Uint64)delay * 1000u / 70u;
+    Uint64 target = SDL_GetTicks() + ms;
 
     IN_StartAck();
